Can be chosen, whether 'no language'-reports on Spreadsheat
The language-id is important for this, so one was temporarily created.
This might give problems, if between viewing reports and
downloading spreadsheat a language was added (by another user)

Because of this it would be better to ommit to have
to times the creating of a new id (for index and for spreadsheat part
in app/controllers/reports_controller.rb.).